+++ tests/test_server.py
@@ -3,6 +3,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 import threading
 import time
 import pytest
+import errno
 import os
 from ssdpy import SSDPServer
 
@@ -29,3 +30,40 @@ def test_server_ipv6():
     time.sleep(0.5)
     server.stopped = True
     server_thread.join()
+
+
+def test_server_invalid_proto():
+    with pytest.raises(ValueError):
+        SSDPServer("test-server", proto="invalid")
+
+
+@pytest.mark.skipif(
+    os.environ.get("TRAVIS") != "true",
+    reason="Not all development environments have a predictable loopback device name",
+)
+def test_server_binds_iface():
+    SSDPServer("test-server", iface=b"lo")
+
+
+def test_server_bind_address_ipv4():
+    SSDPServer("test-server", address="127.0.0.1")
+
+
+@pytest.mark.skipif(
+    os.environ.get("TRAVIS") == "true",
+    reason="IPv6 testing is broken in Travis-CI, see https://github.com/travis-ci/travis-ci/issues/8361",
+)
+def test_server_bind_address_ipv6():
+    SSDPServer("test-server", address="::1", proto="ipv6")
+
+
+@pytest.mark.skipif(
+    os.environ.get("TRAVIS") == "true",
+    reason="IPv6 testing is broken in Travis-CI, see https://github.com/travis-ci/travis-ci/issues/8361",
+)
+def test_server_bind_address_and_iface_ipv6():
+    try:
+        SSDPServer("test-server", address="::1", proto="ipv6", iface=b"lo")
+    except OSError as e:
+        if e.errno != errno.ENOPROTOOPT:  # Protocol not supported
+            raise