+++ CHANGES.md
@@ -3,6 +3,7 @@
 - Fixed an issue where NOTIFY messages were not conforming to UPnP (thanks @hotab)
 - Fixed an issue where `ssdpy.client.discover()` was using wrong syntax.
 - Changed the exception raised by `ssdpy.compat.if_nametoindex()` to be the same as in Python 3 (OSError).
+- Added tests for `ssdpy.client`, `ssdpy.compat` and created more tests for `ssdpy.server` to increase coverage.
 
 ### 0.3.0
 (2020-08-10)